# Avoid O(N^2) in VALUES with ordinals grouping