--- a/x-pack/plugin/esql/compute/src/main/generated-src/org/elasticsearch/compute/aggregation/ValuesDoubleAggregator.java
+++ b/x-pack/plugin/esql/compute/src/main/generated-src/org/elasticsearch/compute/aggregation/ValuesDoubleAggregator.java
@@ -119,6 +119,8 @@ class ValuesDoubleAggregator {
      * Values are collected in a hash. Iterating over them in order (row by row) to build the output,
      * or merging with other state, can be expensive. To optimize this, we build a sorted structure once,
      * and then use it to iterate over the values in order.
+     *
+     * @param ids positions of the {@link GroupingState#values} to read.
      */
     private record Sorted(Releasable releasable, int[] counts, int[] ids) implements Releasable {
         @Override